+++ src/main/java/com/allrounders/goalkeeper/dto/goal/HashtagDTO.java
@@ -1,7 +1,6 @@
-package com.allrounders.goalkeeper.dto;
+package com.allrounders.goalkeeper.dto.goal;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,11 +10,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class GoalJoinMemberDTO {
-
-    @NotNull
-    private Long memberId;
+public class HashtagDTO {
 
     @NotBlank
-    private String nickname;
+    private String tagName;
 }
