+++ src/main/java/com/allrounders/goalkeeper/dto/member/RankDTO.java
@@ -1,4 +1,4 @@
-package com.allrounders.goalkeeper.dto;
+package com.allrounders.goalkeeper.dto.member;
 
 
 import lombok.AllArgsConstructor;
